updated I80F48 to u64 in new log structs

+++ mango-logs/src/lib.rs
@@ -193,7 +193,7 @@ pub struct WithdrawLog {
     pub mango_account: Pubkey,
     pub owner: Pubkey,
     pub token_index: u64,
-    pub quantity: i128,  // I80F48
+    pub quantity: u64,
 }
 
 #[event]
@@ -202,12 +202,12 @@ pub struct DepositLog {
     pub mango_account: Pubkey,
     pub owner: Pubkey,
     pub token_index: u64,
-    pub quantity: i128,  // I80F48
+    pub quantity: u64,
 }
 
 #[event]
 pub struct RedeemMngoLog {
     pub mango_group: Pubkey,
     pub mango_account: Pubkey,
-    pub redeemed_mngo: i128,  // I80F48
+    pub redeemed_mngo: u64,
 }
